move pulse logic to shared
