+++ Content.Server/Clothing/MagbootsSystem.cs
@@ -1,5 +1,6 @@
 using Content.Server.Atmos.Components;
 using Content.Shared.Alert;
+using Content.Shared.Atmos.Components;
 using Content.Shared.Clothing;
 using Content.Shared.Inventory.Events;
